+++ packages/react/src/components/Form.tsx
@@ -68,8 +68,8 @@ export function Form<
 }: FormProps<TFieldValues, TResponseData>): JSX.Element {
   return (
     <form
-      {...props}
       noValidate
+      {...props}
       ref={(element) => (form.element.value = element)}
       onSubmit={async (event) => {
         // Prevent default behavior of browser