+++ packages/react/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Allow `noValidate` prop of `Form` to be overwritten (issue #92)
+
 ## v0.4.1 (June 05, 2023)
 
 - Fix start value bug when using radio inputs (issue #83)